Move toolchain related classes to their own package
