+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/jvm/internal/DefaultJvmTestSuite.java
@@ -30,6 +30,13 @@ import org.gradle.api.plugins.JvmTestSuitePlugin;
 import org.gradle.api.plugins.jvm.JvmComponentDependencies;
 import org.gradle.api.plugins.jvm.JvmTestSuite;
 import org.gradle.api.plugins.jvm.JvmTestSuiteTarget;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.JUnit4TestToolchain;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.JUnitJupiterToolchain;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.JVMTestToolchain;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.KotlinTestToolchain;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.LegacyJUnit4TestToolchain;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.SpockToolchain;
+import org.gradle.api.plugins.jvm.internal.testing.toolchains.TestNGToolchain;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.provider.ProviderFactory;
