+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/jvm/internal/testing/toolchains/AbstractJVMTestToolchain.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.plugins.jvm.internal.testing.toolchains;
+
+import org.gradle.api.internal.tasks.testing.TestFramework;
+import org.gradle.api.tasks.testing.Test;
+
+abstract public class AbstractJVMTestToolchain<T extends JVMTestToolchain.Parameters> implements JVMTestToolchain<T> {
+    private TestFramework testFramework;
+
+    @Override
+    public TestFramework createTestFramework(Test task) {
+        if (testFramework == null) {
+            testFramework = initializeTestFramework(task);
+        }
+
+        return testFramework;
+    }
+
+    abstract protected TestFramework initializeTestFramework(Test task);
+}
