+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/jvm/internal/testing/toolchains/package-info.java
@@ -14,22 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.plugins.jvm.internal;
+@NonNullApi
+package org.gradle.api.plugins.jvm.internal.testing.toolchains;
 
-import org.gradle.api.internal.tasks.testing.TestFramework;
-import org.gradle.api.tasks.testing.Test;
-
-abstract public class AbstractJVMTestToolchain<T extends JVMTestToolchain.Parameters> implements JVMTestToolchain<T> {
-    private TestFramework testFramework;
-
-    @Override
-    public TestFramework createTestFramework(Test task) {
-        if (testFramework == null) {
-            testFramework = initializeTestFramework(task);
-        }
-
-        return testFramework;
-    }
-
-    abstract protected TestFramework initializeTestFramework(Test task);
-}
+import org.gradle.api.NonNullApi;
